- codesnippets: non-windows build fix
git-svn-id: https://svn.code.sf.net/p/codeblocks/code/trunk@10052 2a5c6006-c6dd-42ca-98ab-0921f2732cef

+++ src/plugins/contrib/codesnippets/codesnippetstreectrl.cpp
@@ -1885,10 +1885,10 @@ void CodeSnippetsTreeCtrl::SaveAllOpenEditors()
 // ----------------------------------------------------------------------------
 {
     // 2014/12/20
-    UINT knt = m_EditorPtrArray.GetCount();
+    size_t knt = m_EditorPtrArray.GetCount();
     if (not knt) return;
 
-    for (UINT ii=0; ii<knt; ++ii)
+    for (size_t ii=0; ii<knt; ++ii)
     {
         EditorBase* eb = m_EditorPtrArray[ii];
